+++ contracts/test/tokens/NonRevertTestToken.sol
@@ -1,30 +1,21 @@
 // SPDX-License-Identifier:MIT
 pragma solidity ^0.6.12;
 
-import '../ERC20Mod.sol';
+import "../ERC20Mod.sol";
 
-contract NonRevertTestToken is ERC20Mod('NonRevert Test Token', 'NRTKN') {
-    function mint(uint256 amount, address to) public {
+contract NonRevertTestToken is ERC20Mod("NonRevert Test Token", "NRTKN") {
+    function mint(uint amount, address to) public {
         _mint(msg.sender, amount);
         _transfer(msg.sender, to, amount);
     }
 
-    function transfer(address recipient, uint256 amount)
-        public
-        virtual
-        override
-        returns (bool success)
-    {
+     function transfer(address recipient, uint256 amount) public virtual override returns (bool success) {
         address sender = _msgSender();
-        if (
-            (sender != address(0)) &&
-            (recipient != address(0)) &&
-            (_balances[sender] >= amount)
-        ) {
-            _balances[sender] = _balances[sender].sub(amount);
-            _balances[recipient] = _balances[recipient].add(amount);
-            emit Transfer(sender, recipient, amount);
-            success = true;
+        if( (sender != address(0)) && (recipient != address(0)) && (_balances[sender] >= amount)){
+                _balances[sender] = _balances[sender].sub(amount);
+                _balances[recipient] = _balances[recipient].add(amount);
+                emit Transfer(sender, recipient, amount);
+                success = true;
         }
     }
 }
